ai-packages/chunk-evaluation-module/requirements.txt
pika==1.3.2
chonkie[semantic]
requests
scikit-learn
arize-phoenix
arize-phoenix-otel
pandas

